+++ src/ewoksppf/tests/test_ppf_workflow21.py
@@ -99,8 +99,10 @@ def submodel21_on_error():
 def workflow21(on_error):
     if on_error:
         submodel21 = submodel21_on_error
+        out1_required = False
     else:
         submodel21 = submodel21_conditions
+        out1_required = None
 
     nodes = [
         {"id": "in", "task_type": "method", "task_identifier": qualname(passthrough)},
@@ -132,6 +134,7 @@ def workflow21(on_error):
         {
             "source": "out1",
             "target": "out",
+            "required": out1_required,
             "data_mapping": [{"source_output": "return_value", "target_input": "a"}],
         },
         {